Move outcome_error_message to header
