+++ tiledb/sm/filesystem/s3.cc
@@ -162,60 +162,6 @@ using namespace tiledb::common;
 
 namespace tiledb::sm {
 
-namespace {
-
-/**
- * Return the exception name and error message from the given outcome object.
- *
- * @tparam R AWS result type
- * @tparam E AWS error type
- * @param outcome Outcome to retrieve error message from
- * @return Error message string
- */
-template <typename R, typename E>
-std::string outcome_error_message(const Aws::Utils::Outcome<R, E>& outcome) {
-  if (outcome.IsSuccess()) {
-    return "Success";
-  }
-
-  auto err = outcome.GetError();
-  Aws::StringStream ss;
-
-  ss << "[Error Type: " << static_cast<int>(err.GetErrorType()) << "]"
-     << " [HTTP Response Code: " << static_cast<int>(err.GetResponseCode())
-     << "]";
-
-  if (!err.GetExceptionName().empty()) {
-    ss << " [Exception: " << err.GetExceptionName() << "]";
-  }
-
-  // For some reason, these symbols are not exposed when building with MINGW
-  // so for now we just disable adding the tags on Windows.
-  if constexpr (!platform::is_os_windows) {
-    if (!err.GetRemoteHostIpAddress().empty()) {
-      ss << " [Remote IP: " << err.GetRemoteHostIpAddress() << "]";
-    }
-
-    if (!err.GetRequestId().empty()) {
-      ss << " [Request ID: " << err.GetRequestId() << "]";
-    }
-  }
-
-  if (err.GetResponseHeaders().size() > 0) {
-    ss << " [Headers:";
-    for (auto&& h : err.GetResponseHeaders()) {
-      ss << " '" << h.first << "' = '" << h.second << "'";
-    }
-    ss << "]";
-  }
-
-  ss << " : " << err.GetMessage();
-
-  return ss.str();
-}
-
-}  // namespace
-
 S3Parameters::Headers S3Parameters::load_headers(const Config& cfg) {
   Headers ret;
   auto iter = ConfigIter(cfg, constants::s3_header_prefix);
